Feat: Implements realtime updates to issues
Implements realtime updates for issues and releases
using Firestore listeners. This eliminates the need for manual
state updates after creating or modifying items, ensuring that
the UI reflects the latest data automatically.

Adds a test button in development mode to easily trigger a
realtime update, which helps in verifying the functionality.
Removes the handling of top level epics, epic should have a parent.
Timestamps all issues that are created

AB#129933